[python] extract converter_funcall.cpp from python_converter.cpp
Move 8 function-call-resolution methods of python_converter into
src/python-frontend/converter/converter_funcall.cpp: get_resolved_value,
resolve_function_call, get_function_constant_return, is_identity_function,
get_function_call, store_call_result, get_return_from_func,
materialize_list_function_call.

Pure code movement — no logic change.
